Add Regexp Matcher
Adds a matcher which takes a string to interpreted as a regular
expression accoring to the [regexp][1] package.

This can be useful when the string might vary depending on factors that
are outside of the testers control (or simply not worth the effort of
taking control over) such as a timestamp.

[1]: https://golang.org/pkg/regexp/